// src/interfaces/syllabusTypes.ts
export interface Syllabus {
  subject: string;
  educational_program: string;
  specialty: string;
  level_of_education: string;
  language: string;
  abstract: string;
  goal: string;
  competencies: string;
  outcomes: string;
  content: string;
  methods: string;
  evaluation: string;
  literature: string;
}